SS 4.0 - Fix QueuedJobService, QJUtils and QueuedJobsTests
